Add files via upload
In this post, I look into Geometric Brownian Motion and how it helps us evaluate risk and return. GBM is often used to explain the movement of time series variables and, in finance, plays a crucial role in modeling the movement of asset prices. As the object of analysis I used the SPY ETF, which replicates the S&P 500 index.

These simulations can help us evaluate the risk of our portfolio using the VaR (Value at Risk) methodology. Furthermore, I showed the impact of a 5% increase in monthly volatility and its impact on the VaR.